Añadidas excepciones para ProductProviderUpdater

+++ src/main/java/ar/edu/ungs/fleet_manager/products/application/update/ProductProviderUpdater.java
@@ -1,13 +1,15 @@
 package ar.edu.ungs.fleet_manager.products.application.update;
 
-import ar.edu.ungs.fleet_manager.products.application.ProductProviderRequest;
 import ar.edu.ungs.fleet_manager.products.domain.Product;
 import ar.edu.ungs.fleet_manager.products.domain.ProductId;
 import ar.edu.ungs.fleet_manager.products.domain.ProductRepository;
 import ar.edu.ungs.fleet_manager.products.domain.services.ProductFinder;
+import ar.edu.ungs.fleet_manager.providers.application.ProviderResponse;
+import ar.edu.ungs.fleet_manager.providers.application.search.ProviderByProductSearcher;
 import ar.edu.ungs.fleet_manager.providers.domain.Provider;
 import ar.edu.ungs.fleet_manager.providers.domain.ProviderId;
 import ar.edu.ungs.fleet_manager.providers.domain.services.ProviderFinder;
+import ar.edu.ungs.fleet_manager.shared.domain.exceptions.InvalidParameterException;
 import org.springframework.stereotype.Component;
 
 
@@ -17,11 +19,13 @@ public class ProductProviderUpdater {
     private final ProductRepository repository;
     private final ProductFinder productFinder;
     private final ProviderFinder providerFinder;
+    private final ProviderByProductSearcher searcher;
 
-    public ProductProviderUpdater(ProductRepository repository, ProductFinder productFinder, ProviderFinder providerFinder) {
+    public ProductProviderUpdater(ProductRepository repository, ProductFinder productFinder, ProviderFinder providerFinder, ProviderByProductSearcher searcher) {
         this.repository = repository;
         this.productFinder = productFinder;
         this.providerFinder = providerFinder;
+        this.searcher = searcher;
     }
 
 
@@ -30,6 +34,14 @@ public class ProductProviderUpdater {
         Provider provider = this.providerFinder.execute(new ProviderId(providerId));
 
 
+        if (this.searcher.execute(product.id().value())
+                .stream()
+                .map(ProviderResponse::id)
+                .anyMatch(provider_id -> provider_id.equals(provider.id().value()))) {
+            throw new InvalidParameterException(String.format(
+                    "The product with ID '%s' already has this provider associated.",
+                    product.id().value()));}
+
         this.repository.addProvider(product.id(),provider.id());
     }
 
@@ -37,6 +49,13 @@ public class ProductProviderUpdater {
         Product product = this.productFinder.execute(new ProductId(productId));
         Provider provider = this.providerFinder.execute(new ProviderId(providerId));
 
+        if (this.searcher.execute(product.id().value())
+                .stream()
+                .map(ProviderResponse::id)
+                .noneMatch(provider_id -> provider_id.equals(provider.id().value()))) {
+            throw new InvalidParameterException(String.format(
+                    "The product '%s' is not provided by the provider '%s'.",
+                    product.id().value(),provider.id().value()));}
 
         this.repository.deleteProvider(product.id(),provider.id());
     }
